Move common.content_models to common.models.content
