+++ hasher-matcher-actioner/hmalib/lambdas/actions/action_performer.py
@@ -8,7 +8,7 @@ from mypy_boto3_dynamodb import DynamoDBServiceResource
 
 from hmalib.common.message_models import ActionMessage
 from hmalib.common.configs.actioner import ActionPerformer
-from hmalib.common.content_models import ActionEvent
+from hmalib.common.models.content import ActionEvent
 from hmalib.common.logging import get_logger
 from hmalib.common import config
