+++ hasher-matcher-actioner/hmalib/lambdas/api/submit.py
@@ -23,7 +23,7 @@ from threatexchange.signal_type.pdq import PdqSignal
 
 from hmalib.lambdas.api.middleware import jsoninator, JSONifiable, DictParseable
 from hmalib.common.content_sources import S3BucketContentSource
-from hmalib.common.content_models import ContentObject, ContentRefType
+from hmalib.common.models.content import ContentObject, ContentRefType
 from hmalib.common.logging import get_logger
 from hmalib.common.message_models import URLSubmissionMessage
 from hmalib.models import PipelineHashRecord
